Moved firebase login logic to LogInAction
Makes more sense there
