+++ SeviBus/src/main/java/com/sloy/sevibus/ui/LoginController.java
@@ -5,9 +5,6 @@ import android.content.Intent;
 import android.support.annotation.NonNull;
 import android.util.Log;
 
-import com.firebase.client.AuthData;
-import com.firebase.client.Firebase;
-import com.firebase.client.FirebaseError;
 import com.google.android.gms.auth.GoogleAuthException;
 import com.google.android.gms.auth.GoogleAuthUtil;
 import com.google.android.gms.auth.api.Auth;
@@ -15,22 +12,14 @@ import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.auth.api.signin.GoogleSignInResult;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.gson.Gson;
 
 import java.io.IOException;
-import java.util.Map;
 
 import rx.Observable;
 import rx.android.schedulers.AndroidSchedulers;
 import rx.schedulers.Schedulers;
 
 public class LoginController implements GoogleApiClient.OnConnectionFailedListener {
-
-    private final Firebase firebase;
-
-    public LoginController(Firebase firebase) {
-        this.firebase = firebase;
-    }
 
     public Intent loginIntent(GoogleApiClient googleApiClient) {
         return Auth.GoogleSignInApi.getSignInIntent(googleApiClient);
@@ -40,16 +29,13 @@ public class LoginController implements GoogleApiClient.OnConnectionFailedListen
         Auth.GoogleSignInApi.signOut(googleApiClient);
     }
 
-    public Observable<SevibusUser> handleSignInResult(Context context, Intent data) {
+    public Observable<String> obtainOAuthTokenFromSignInResult(Context context, Intent data) {
         return Observable.just(data)
           .map(Auth.GoogleSignInApi::getSignInResultFromIntent)
           .flatMap(this::successOrFail)
           .map(GoogleSignInResult::getSignInAccount)
           .map(GoogleSignInAccount::getEmail)
-          .flatMap(email -> getOauthToken(context, email))
-          .flatMap(this::autenticateFirebase)
-          .map(this::createSevibusUser);
-
+          .flatMap(email -> getOauthToken(context, email));
     }
 
     private Observable<String> getOauthToken(Context context, String email) {
@@ -65,40 +51,6 @@ public class LoginController implements GoogleApiClient.OnConnectionFailedListen
           .subscribeOn(Schedulers.io())
           .observeOn(AndroidSchedulers.mainThread());
 
-    }
-
-    private Observable<AuthData> autenticateFirebase(String oauthToken) {
-        return Observable.create(subscriber -> firebase.authWithOAuthToken("google", oauthToken, new Firebase.AuthResultHandler() {
-            @Override
-            public void onAuthenticated(AuthData authData) {
-                Log.d("FireLogin", "onAuthenticated!");
-                if (!subscriber.isUnsubscribed()) {
-                    subscriber.onNext(authData);
-                    subscriber.onCompleted();
-                }
-            }
-
-            @Override
-            public void onAuthenticationError(FirebaseError firebaseError) {
-                Log.w("FireLogin", "onAuthenticationError :(!");
-                Log.w("FireLogin", "message: " + firebaseError.getMessage());
-                if (!subscriber.isUnsubscribed()) {
-                    subscriber.onError(firebaseError.toException());
-                }
-            }
-        }));
-    }
-
-    private SevibusUser createSevibusUser(AuthData authData) {
-        Log.d("authData", new Gson().toJson(authData));
-        Map<String, Object> data = authData.getProviderData();
-        SevibusUser user = new SevibusUser();
-        user.setId((String) authData.getAuth().get("uid"));
-        user.setEmail((String) data.get("email"));
-        user.setName((String) data.get("displayName"));
-        user.setPhotoUrl((String) data.get("profileImageURL"));
-        user.setOauthToken((String) data.get("accessToken"));
-        return user;
     }
 
     @Override
